chore: Move utils/meta
