+++ pkg/execution/stores/activejobstore/store_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/furiko-io/furiko/pkg/execution/util/jobconfig"
 	furiko "github.com/furiko-io/furiko/pkg/generated/clientset/versioned"
 	"github.com/furiko-io/furiko/pkg/runtime/controllercontext/mock"
-	"github.com/furiko-io/furiko/pkg/utils/k8sutils"
+	"github.com/furiko-io/furiko/pkg/utils/meta"
 )
 
 const (
@@ -173,7 +173,7 @@ func TestStore(t *testing.T) {
 		go func(i int) {
 			defer wg.Done()
 			newRj := rj.DeepCopy()
-			k8sutils.SetAnnotation(newRj, "index", strconv.Itoa(i))
+			meta.SetAnnotation(newRj, "index", strconv.Itoa(i))
 			newRj.Status.Phase = execution.JobSucceeded
 			updatedRj, err := updateJob(ctx, client, newRj)
 			assert.NoError(t, err)
